Dividing up the various p2p types and interfaces
