+++ contract_scripts_test.go
@@ -6,7 +6,7 @@ import (
 	"strings"
 	"testing"
 
-	tezosprotocol "github.com/anchorageoss/tezosprotocol/v2"
+	"github.com/anchorageoss/tezosprotocol/v2"
 	"github.com/stretchr/testify/require"
 )
