+++ keys_test.go
@@ -10,7 +10,7 @@ import (
 	"encoding/hex"
 	"testing"
 
-	tezosprotocol "github.com/anchorageoss/tezosprotocol/v2"
+	"github.com/anchorageoss/tezosprotocol/v2"
 	"github.com/btcsuite/btcd/btcec"
 	"github.com/stretchr/testify/require"
 	"golang.org/x/crypto/ed25519"
